ath11k: Avoid possible crash when radio is disabled on bootup
When a particular radio is skipped during ahb_probe, the
platform drv data is currently not set, and during ahb remove
we access the drv data directly without checking if its valid.

This causes kernel crash due to invalid memory access.
Hence, set the platform drv data to NULL when a radio is skipped
and check for the same during ahb_remove().

   40.532126] Unable to handle kernel paging request at virtual address 0000d7b4
[   40.532155] pgd = 952ac000
[   40.538407] [0000d7b4] *pgd=00000000
[   40.544568] Internal error: Oops: 5 [#1] PREEMPT SMP ARM
<snip>
   40.810363] CPU: 0 PID: 1947 Comm: rmmod Not tainted 4.4.60 #1
[   40.810712] Hardware name: Generic DT based system
[   40.816442] task: 9cd8a300 ti: 9528a000 task.ti: 9528a000
[   40.821225] PC is at ath11k_ahb_remove+0x10/0xf8 [ath11k_ahb]
[   40.826691] LR is at platform_drv_remove+0x24/0x3c

Fixes: I78e571108 ("mac80211: Package upgrade")

Change-Id: I11ebf3798eee2b64f5aab85c5453f1b1f4d29fc8
Signed-off-by: Sriram R <srirrama@codeaurora.org>