ath11k: disabling auto scan for the mesh mode in wpa_supplicant
noscan disables the scanning of bandwidth availablity check
for 40Mhz in hostapd. When in mesh mode with encryp wpa_supplicant
is used, here 'disable_40mhz_scan' is the term used for noscan in
wpa_supplicant.

So setting 'disable_40mhz_scan' in wpa supp conf when noscan is enabled
in wireless config.

Signed-off-by: Karthikeyan Kathirvel <kathirve@codeaurora.org>
Change-Id: I3e7fcc26a475881e6a3a60bad2eaa18ad4647777