ath11k: fix exception enum and exception msg value
There is a mismatch in exception enum type between nss driver and ath11k
due to which exception is not properly working and also exception value
was not properly assigned with unsigned int instead it was assigned with
an address of that variable.

Corrected the WIFI_MESH_EXT_DATA_PKT_TYPE_EXCEPTION to
NSS_WIFI_VDEV_EXT_DATA_PKT_TYPE_MESH and exception value.

Signed-off-by: Karthikeyan Kathirvel <kathirve@codeaurora.org>
Change-Id: Ied492f5bbd53f549b792d73478b13c48260d76ce