ath10k: validate first subframe of amsdu before processing the list
Validate amsdu list by checking if the first frame has a valid
DA. In certain scenarios a proper msdu can be received as an amsdu when
the A-MSDU present bit gets flipped and corresponding FCS corrected.
Since this bit is unauthenticated the HW crypto engine can pass the frame
to driver without any error indication. This could result in processing
unintended subframes collected in the amsdu list.

Comparing the non-aggregated msdu and an amsdu, the fields of the first
subframe DA matches with the llc/snap header fields of normal msdu.
In order to avoid processing such frames, add a validation to
filter such amsdu frames where the first subframe header DA matches
with the llc/snap header pattern.

Tested-on: QCA6174 hw3.2 PCI WLAN.RM.4.4.1-00110-QCARMSWP-1

Change-Id: Ifc368799aa7be2c8c23b422027c906d1a456ddb6
Signed-off-by: Sriram R <srirrama@codeaurora.org>
Signed-off-by: Wen Gong <wgong@codeaurora.org>
Signed-off-by: Ronghua Zeng <ronghuaz@codeaurora.org>