ath10k: add PN replay protection for fragmentation for PCIe
Recently the PN replay check for NON-fragmentaion is finished in
firmware, but for fragmentaion, it does not have PN replay check
in ath10k for PCIe. It has function ieee80211_rx_h_defragment in
mac80211 which do PN replay check for fragmentaion, but it does
not match QCA6174, it check condition ieee80211_has_protected, it
is not match QCA6174.

This patch is to get the PN number for CCMP and TKIP of fragmentation
and drop the fragmentation which PN number is not correct. It only
effect QCA6174/QCA6377 PCIe.

Tested-on: QCA6174 hw3.2 PCI WLAN.RM.4.4.1-00110-QCARMSWP-1

Change-Id: Ibe36ffa1aa7e72b67845db3b405d3694e75d44f4
Signed-off-by: Wen Gong <wgong@codeaurora.org>
Signed-off-by: Ronghua Zeng <ronghuaz@codeaurora.org>