ath10k: set flag for tkip michael mic verifyi for PCIe
Recently it does not have tkip michael mic verify for PCIe, it has
ieee80211_rx_h_michael_mic_verify in mac80211 which do the verify,
but it has some check for the flag of ieee80211_rx_status, and ath10k
do not meet its check.

This patch is set the flag property to meet check of mac80211, this
fixs CVE-2020-26141. It do the same as ath10k_htt_rx_proc_rx_ind_hl
for SDIO which passed mic verify case. It only effect QCA6174/QCA9377
PCIe.

Tested-on: QCA6174 hw3.2 PCI WLAN.RM.4.4.1-00110-QCARMSWP-1

Change-Id: I30b5b57911cbb1c1a680cdc1e780babe0afb9ef5
Signed-off-by: Wen Gong <wgong@codeaurora.org>
Signed-off-by: Ronghua Zeng <ronghuaz@codeaurora.org>