ath11k: Fix ap vlan object double add to ap vlans list
Currently the ap vlan arvif object maintained in the AP's vlan arvif
list gets corrupted when we add the same vlan object which could
happen when a WDS sta reassociates where the same vif object
would be associated with a newly created sta. This double add
causes list corruption leading to hangs or invalid data access
when one of the object gets deleted but other object in the
list gets accessed during nss vdev cleanup of AP vap.

Hence add appropriate checks to see if the vif object
is already added to the list, also use the checks to
avoid delete if the object was never added during
remove interface calls.

Also, when a new peer is associated with already existing
vif object, we need to send WDS reconfigure for the
associated peer and set 4addr allow in NSS FW to
resume traffic for the peer.

Change-Id: I32bbc1d8d39d4f281439b0578bb2771fb462123b
Signed-off-by: Sriram R <srirrama@codeaurora.org>