cfg80211: mitigate A-MSDU aggregation issue
Mitigate A-MSDU injection issues (CVE-2020-24588) by detecting if the
destination address of a subframe equals an RFC1042 (i.e., LLC/SNAP)
header, and if so dropping the complete A-MSDU frame. This mitigates
known issues, although new (unknown) aggregation-based issues may
remain possible.

This defense works because in A-MSDU aggregation injection issues, a
normal encrypted Wi-Fi frame is turned into an A-MSDU frame. This means
the first 6 bytes of the first A-MSDU subframe correspond to an RFC1042
header. In other words, the destination MAC address of the first A-MSDU
subframe contains the start of an RFC1042 header during an aggregation
issue. We can detect this and thereby prevent this specific issue.
For details, see Section 7.2 of "Fragment and Forge: Breaking Wi-Fi
Through Frame Aggregation and Fragmentation".

Note that for kernel 4.9 and above this patch depends on "mac80211:
properly handle A-MSDUs that start with a rfc1042 header". Otherwise
this patch has no impact and issues will remain possible.

Cherry-picked from
https://source.codeaurora.org/quic/qsdk/kvalo/ath/commit/?h=kernel-ath/master&id=2b8a1fee3488c602aca8bea004a087e60806a5cf

Git-Commit: 2b8a1fee3488c602aca8bea004a087e60806a5cf
Git-Repo: https://source.codeaurora.org/quic/qsdk/kvalo/ath

Signed-off-by: Mathy Vanhoef <Mathy.Vanhoef@kuleuven.be>
Link: https://lore.kernel.org/r/20210511200110.25d93176ddaf.
	      I9e265b597f2cd23eb44573f35b625947b386a9de@changeid
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: Sriram R <srirrama@codeaurora.org>
Change-Id: I3321a9e0a0a04989b832d8ddc1050fc9bdd8439e