mac80211: drop A-MSDUs on old ciphers
With old ciphers (WEP and TKIP) we shouldn't be using A-MSDUs
since A-MSDUs are only supported if we know that they are, and
the only practical way for that is HT support which doesn't
support old ciphers.

However, we would normally accept them anyway. Since we check
the MMIC before deaggregating A-MSDUs, and the A-MSDU bit in
the QoS header is not protected in TKIP (or WEP), this enables
attacks similar to CVE-2020-24588. To prevent that, drop A-MSDUs
completely with old ciphers.

Cherry-picked from
https://source.codeaurora.org/quic/qsdk/kvalo/ath/commit/?h=kernel-ath/master&id=270032a2a9c4535799736142e1e7c413ca7b836e

Git-Commit: 270032a2a9c4535799736142e1e7c413ca7b836e
Git-Repo: https://source.codeaurora.org/quic/qsdk/kvalo/ath

Link: https://lore.kernel.org/r/20210511200110.076543300172.
	      I548e6e71f1ee9cad4b9a37bf212ae7db723587aa@changeid

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: Sriram R <srirrama@codeaurora.org>
Change-Id: I52c817765bae459214aadd61f13cb3c3b340c829