ath11k: enable reo status poll timer for qcn6122
During long run tests it is observed that the reo status
ring is not reaped as interrupts are not received properly

Enable reo status poll timer as in IPQ5018 to avoid this
issue.

Also, ensure the timer flag is reset after dp cleanup.

Note that there is no performance impact because of
these timers since they are activated only during reo
command usage which happens only during reo queue setup
and updates.

Change-Id: Ieb769698a7976a339e0db07c674895eb6a9f58cb
Signed-off-by: Sriram R <srirrama@codeaurora.org>