ath11k : fix max peer count in ipq5018 and qcn6122
When fw receives the peer_create wmi cmd for 129th peer, the peer pool in
fw gets exhausted and it leads to Q6 crash in ipq5018 and qcn6122.

Currently the maximum number of peers is set as 512 and the maximum
number of vdevs is set as (16+1) for all the platforms with 1GB profile.

But the maximum number of peers is 128 and the maximum number of
vdevs is (8+1) for ipq5018 and qcn6122 with 1GB profile.

Signed-off-by: Hari Chandrakanthan <haric@codeaurora.org>
Change-Id: I8102059e35db0d91e85b66f48e4d8c8b773f0464