ath11k: resetting the mpp num cnt before dump request
mpp_dump_num_entries is not reset properly this causing the number of
entries to be grown larger for every request of mpp dump cmd hence
kernel is not able to allocate the requested larger memory.

Properly reset the mpp_dump_num_entries and made sync with the
mpp_timer_cb to not disturb each others request.

Change-Id: Ieef3f563275f2798bdb33a6d808945616c426c10
Signed-off-by: Karthikeyan Kathirvel <kathirve@codeaurora.org>