ath11k: failed to update mpath entry
If there is a mpath update which has a better nexthop to reach
the destination for the existing mpath this is not updating the
new nexthop for existing path to NSS.

As NSS requires both dest and nexthop address to pick the specific
mpath, in result NSS returns path not found error to the host.

Sending old_next_hop address to NSS only if there is a path change based
on this NSS should update its mpath table.

Change-Id: Iea09d797356507a48a9bb6024b71a36a67dbda21
Signed-off-by: Karthikeyan Kathirvel <kathirve@codeaurora.org>