ath11k: Restrict the duplicate peer creation on SOC level
Rhash address table is created radio specific which leads to mislead
due to partial table is maintained in SOC level and partial maintained
in radio level. So to avoid this mislead, we moved all peer table in
SOC level and also avoid the duplicate peer in SOC level since no
valid usecase identifed.

Fixes: Ic1de3048e29c (ath11k: fix peer delete for already deleted peer)

Change-Id: I72c0b184c2c2a1db61142b7997eb4a40e254d365
Signed-off-by: Karthikeyan Periyasamy <periyasa@codeaurora.org>