mac80211: Add workaround to send 4addr packet from repeater
Once the repeater sta associates to AP, it has to send some 4addr
packets to AP. It will help creating AP-VLAN entry in AP for
the repeater. So, add workaround to send some arp broadcast packets
which will be in 4addr. This is sent once STA interface status
becomes COMPLETED. This is applicable for AP+STA mode.

Change-Id: Idb7657f3cfc28276fc91e92d5488c04940f7162c
Signed-off-by: Seevalamuthu Mariappan <seevalam@codeaurora.org>