ath11k: Add packet latency histogram support
Latency of a packet is measured from packet being enqueued
to the driver's transmit queue, to frame has been successfully
ack’ed by the receiver. Enqueue time is added in tx info and
difference in time is calculated once the packet is acked by
the receiver to get latency. Latency is calculated in powers
of 2 and number of packets that took a particular time is
incremented in the corresponding index and displayed via debugfs
These stats has to be enabled through enable_tx_latency_stats debugfs
as shown below

$ echo 1 > /sys/kernel/debug/ieee80211/phy1/enable_tx_latency_stats

$ cat /sys/kernel/debug/ath11k/ipq8074 hw2.0/mac0/tx_delay_histogram/AC_VO+
TX delay histogram(ms)
[   0 -  1): 1 [  1 -  2):38 [  2 -  3):46 [  3 -  4):17 [  4 -   6):19
[   6 -  8):10 [  8 - 12): 5 [ 12 - 16): 1 [ 16 - 24): 3 [ 24 -  32): 0
[  32 - 48): 0 [ 48 - 64): 0 [ 64 - 96): 0 [ 96 -128): 0 [128 - 192): 0
[ 192 -256): 0 [256 -384): 0 [384 -512): 0 [512 -768): 0 [768 -1024): 0
[1024 -inf): 0

Change-Id: I1fcc41ac659e4e2e1ed71461b49fe4623749fe8b
Signed-off-by: Sowmiya Sree Elavalagan <quic_ssreeela@quicinc.com>