ath12k: fix workqueue not getting destroyed after rmmod
This change is to destroy the workqueues after the modules
are removed from kernel.Earlier, the workqueues were
not destroyed and left as it is.

If workqueues are not destroyed, it could lead to kernel
memory scarcity in a longer run. This could affect self and
other drivers workability as well.

Change-Id: I480e61e65d1763e71fc9094bfa762e3d829e61fa
Signed-off-by: Aditya Kumar Singh <adisi@codeaurora.org>
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>