ath11k: fix workqueue not getting destroyed after rmmod
This change is to destroy the workqueues after the modules
are removed from kernel. Earlier, the workqueues were
not destroyed and left as it is.

If workqueues are not destroyed, it could lead to kernel
memory scarcity in a longer run. This could affect self and
other drivers workability as well.

Change-Id: I0c0a880f763a6cfcdcc7cde6da7119a225bc4f56
Signed-off-by: Aditya Kumar Singh <adisi@codeaurora.org>