mac80211: Add delay in ap+sta mode script
There is an infinite while loop running in apsta_mode.sh.
This causes CPU overload. So, add delay of 100ms in while
loop. This reduces CPU load upto 60-70%.

Also, when interface is deleted, the script is still running
in background. When interface is created again, new instance
for the script is getting created. This again causes CPU overhead.
Hence, kill the process when bringing down the interface.

Change-Id: I0638c50dacdd2cb6a5ac53e928d40ecba3dccf3b
Signed-off-by: Seevalamuthu Mariappan <seevalam@codeaurora.org>