ath11k: update rx rate stats for all peers in PPDU for MU reception
Rx rate stats are updated to peer nss stats in monitor
status tlv parsing which will be updated to upper layer
for station info.
Currently rx stats update for all users in MU PPDU Reception is
not done. Fix the code to loop through all users in case of
MU reception and update stats for all peers.

Fixed MCS value for HT preamble, which assumes mcs
is exclusive of the nss fields, whereas the mcs
from ppdu_info is received including nss info
and this can be used directly.
Removed num_mdsu which is assigned but never used.

Change-Id: I3b2a213ed6f36e7c552b7e5d336e6481228ca739
Co-developed-by: Sriram R <srirrama@codeaurora.org
Signed-off-by: Sriram R <srirrama@codeaurora.org>
Signed-off-by: Rameshkumar Sundaram <ramess@codeaurora.org>