ath11k: clear the keys properly when DISABLE_KEY
DISABLE_KEY sets the key_len to 0, firmware will not delete the keys if
key_len is 0. Changing from security mode to open mode will cause mcast
to be still encrypted without vdev restart.

Set the proper key_len for DISABLE_KEY cmd to clear the keys in
firmware.

Without vdev restart sta state is not moved to authorized in firmware so
this cause unicast drop in firmware.

Since rsnie_present and wpaie_present field holds the old state when
moved from secure mode to open mode, this clears the wmi authorize flag,
so peer is not authorized in this case. Fix it by reseting rsnie_present
and wpaie_present when the IE is not present.

Change-Id: Ib6de7d1f6eee1340bdc7fcff156c4930f3d891b8
Signed-off-by: Karthikeyan Kathirvel <kathirve@codeaurora.org>