ath11k: update all 6G reg rules in sband
Based on the power modes in 6G band, there can be total of 9
possible channel arrays in sband. This patch updates the 6G
sband with all possible 9 modes. The sband will be updated
with default values since we dont know the power mode during
sband setup (initial boot).

Once the FW sends the reg rules and regd is built, cfg80211
will update the sband as per the rules. To aid cfg80211 to
distinguish between the power mode, the ath11k updates the
power mode as well while building regulatory domain.

Change-Id: I37c9955c5ba2d7d467ff298facf28ce75709ec64
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>