ath11k: Use array read for 'caldb-addr' dts entry
'caldb-addr' dts entry is an array entry for internal
radios in dts files. So, use 'of_property_read_u32_array'
to read the same. For pci nodes, this entry is used as u32 in
dts files. So, use 'of_property_read_u32' to read caldb
addr if array read failed.

Change-Id: I932ee052ba050d4cdafb3aea76b34289daa9d1e7
Signed-off-by: Seevalamuthu Mariappan <quic_seevalam@quicinc.com>