ath11k: sequence disassoc frame and clear key
During disassociation, the disassoc frame and clear key wmi cmd
are sent to fw back to back.
Due to this, fw clears the key and then processes the disassoc frame.
When PMF is enabled, this leads to sending the disassoc frame
without proper encryption. Hence, send the clear key wmi cmd
to fw only after receiving the tx completion of disassoc frame.

Signed-off-by: Hari Chandrakanthan <quic_haric@quicinc.com>
Change-Id: Icde5ebc3adb778bfb2267a2c951ad984f2bc4578