ath11k: inceasing the max num of ssid/bssid for prb req
The maximum number of SSIDs in a for active probe requests is currently
reported as 16 (WLAN_SCAN_PARAMS_MAX_SSID) when registering the driver.
The scan_req_params structure only has the capacity to hold 10 SSIDs.
This leads to a buffer overflow which can be triggered from
wpa_supplicant in userspace. When copying the SSIDs into the
scan_req_params
structure in the ath11k_mac_op_hw_scan route, it can overwrite the
extraie pointer.

Firmware supports 16 ssid * 4 bssid, for each ssid 4 bssid combo probe
request will be sent, so totally 64 probe requests supported. So
increasing both max ssid and bssid to 16 and 4 respectively.

Change-Id: Ie102ab2900b7d0834bfafb9b9242986badd363d9
Signed-off-by: Karthikeyan Kathirvel <quic_kathirve@quicinc.com>