ath11k: Fix ssid in 6G scan request to firmware.
On trying to connect ath11k station to a hidden ssid, the
probe request sent had a wrong ssid. This is because the
ssid field wmi scan start command was corrupted due to
array out of bound while filling channel list for 6g.
Initially function was implemented to handle 5G channels
which were limited, currently on using same for 6G channel
array went out of bound due to high number of channels in
6G. Extended the array index to accommodate 6G channels as well.

Change-Id: I35c91d153447edeaba2aebd50b865aeed34a3f21
Signed-off-by: Sowmiya Sree Elavalagan <quic_ssreeela@quicinc.com>