mac80211: loop over phy in natural order
The loop is reading files in the below order as it is sorting in
alphanumerical order
/sys/class/ieee80211/phy10
/sys/class/ieee80211/phy11
/sys/class/ieee80211/phy9

Due to this "wifi config > /etc/config/wireless" cmd wrongly shuffling
the configs between phy10 and phy9 so radio cannot be turned on with
invalid config. This is the case for 99,100,101...999,1000,1001...etc
sequence.

Sort the order of phy# in natural order to fix this issue, the correct
order in natural sort after fixing it
/sys/class/ieee80211/phy9
/sys/class/ieee80211/phy10
/sys/class/ieee80211/phy11

Change-Id: I1aa7e430262daa6f58462c9c2c2c3cc594be57f4
Signed-off-by: Karthikeyan Kathirvel <quic_kathirve@quicinc.com>