ath11k: dtim count configuration for non-tx VAPs
A temporary work-around was added to set DTIM count for non-transmitting
AP interfaces to 0 until those are brought up when multiple BSSID
feature is enabled.

It is not needed anymore as firmware has made a wmi_resource_config
field available (flags2, bit 9) to do the DTIM count calculation
in the target itself.

Set the resource configuration and remove the work-around.

Change-Id: Ic207cee67d2ce317f7825ee5e7087325f0b50f7c
Signed-off-by: Aloka Dixit <quic_alokad@quicinc.com>