ath12k: Rename ARRAY_TO_STRING to PRINT_ARRAY_TO_BUF
Rename the macro ARRAY_TO_STRING to describe it's
functionality more vividly. Increment the buffer
length within the macro to ease increment within
APIs.

Remove the htt_stats fixed size array usage.
HTT Stats DebugFS interface a single large buffer
that contains the stats must be provided to the
DebugFS infrastructure. With current implementation
for each class of stats, the stats are first
formatted in a local on-stack buffer, and then the
local buffer is copied to the large DebugFS buffer.

When for a given class, the formatted stats exceeds
the size of the on-stack buffer the stats for this
class is truncated. In addition, this logic is
inefficient since it introduces an unnecessary
memory copy.

Write the formatted data directly into the large
DebugFS buffer to resolve the issue.

Change-Id: I169cb334f0f4df3ce1f4505e031b4f59122631d8
Signed-off-by: Dinesh Karthikeyan <quic_dinek@quicinc.com>