ath12k: Add single wiphy registration support
Single wiphy registration support to enable MLO.

With various design options in wiphy registration and interface creations
for MLO, it is extremely difficult to support some of the requirements
in MLD. So for there is a separate wiphy registered with cfg80211 for every
radio (phy) irrespective of whether those are part of the same NIC or not.
So we have to move away from multi-wiphy model to support MLO across multiple
radio/SOC.

The proposal is to combine all the radios under a single wiphy as separate bands
and handle pdev/soc access difference in the driver. In MLD, there will be
only one interface exposed to user space, driver abstraction layer handles
communications with the fw/hw and vice-versa. This avoids significant hostapd,
cfg80211 and mac80211 overheads especially in MLO capability advertisement and
across wiphy/hw access.

Change-Id: I7fe6c1231921c1a87bde48a2c352f78441bb9192
Signed-off-by: Karthikeyan Periyasamy <quic_periyasa@quicinc.com>