ath12k: add wait operation for tx management packets for flush from mac80211
Transmission of management packets are done in a work queue.
Sometimes the workqueue does not finish Tx immediately, then
it lead after the next step of vdev delete finished, it start
to send the management packet to firmware and lead firmware crash.

ieee80211_set_disassoc() have logic of ieee80211_flush_queues()
after it send_deauth_disassoc() to ath12k, its purpose is make
sure the deauth was actually sent, so it need to change ath12k
to match the purpose of mac80211.

To address these issue wait for Tx management as well as Tx
data packets.

Change-Id: I3340c38e484674eea6fae19ae334f2e7f412cf79
Signed-off-by: Wen Gong <quic_wgong@quicinc.com>
Signed-off-by: Karthik M <quic_karm@quicinc.com>