ath11k: fix contiguous qmi memory allocation for pci based devices
From DT the property "base-addr" along with some other properties of memory
regions which contains addresses of those regions is removed and
"memory-region" property is added, to remove the memory address handling
in DT according to upstream.

This "base-addr" property is used to denote whether fixed memory region
assignment or dynamic memory allocation is to be used.
Since this is removed, making use of "memory-region" property to make
this decision. In FTM mode, only with fixed memory region assignment,
the correct power values are returned during the Tx calibration test cases.

Fixes: 3bd93de22933 ("ath11k: add fw_mem_mode hw_param and qmi base memory support")
Change-Id: Id6534fab670e70d60060a489aa00bfcbb2f7102d
Signed-off-by: Nagarajan Maran <quic_nmaran@quicinc.com>