ath12k: update peer assoc flag in WMI
When STA associates in 320MHZ, since in peer assoc command appropriate
peer assoc flags are not set, FW decides that the STA got associated in 20MHz.
This in turn affects the HTT stats counter with respect to bw.

Updated the peer assoc flag accordingly and verified the HTT stats for 320MHz.

Change-Id: I1e97277055686d6e5a38effe9bf34a52ba8dc8f4
Signed-off-by: Aaradhana Sahu <quic_aarasahu@quicinc.com>