ath11k: Fix qmi_msg_handler data structure initialization
qmi_msg_handler is required to be null terminated by QMI module.
There might be a case where a handler for a msg id is not present in the
handlers array which can lead to infinite loop while searching the handler
and therefore out of bound access in qmi_invoke_handler().
 Hence update the initialization in qmi_msg_handler data structure.

Change-Id: Ida98d2018ab0bcecfdcef941e65e6d1e014b486d
Signed-off-by: Rahul Bhattacharjee <quic_rbhattac@quicinc.com>