mac80211: Fix dynamic vlan warning with monitor interface restart
When monitor interface restarts, in nss offload disabled
case, the encap and decap offload flags are removed
from all the interfaces in that phy#.

However when dynamic VLAN and monitor interfaces are
created in the same phy#, these flags are not updated
correctly, due to which warning calltrace is observed.

Add condition check to update the correct flags in
dynamic VLAN case.

Change-Id: I8be0c2a6b68c03e2cbc839179cc8c6b635e50ffe
Signed-off-by: Nagarajan Maran <quic_nmaran@quicinc.com>