ath11k: fix monitor vdev creation with firmware recovery
During firmware recovery, the monitor interface is not
getting created in the driver and firmware since
the respective flags are not updated properly.

So after firmware recovery is successful, when monitor
interface is brought down manually, firmware assertion
is observed since we are trying to bring down the
interface which is not yet created in the firmware.

Fix this by updating the monitor flags properly per
phy#, during firmware recovery.

Change-Id: Id88ef3105b03ec61b88a4bae49315bf64f2542ee
Signed-off-by: Nagarajan Maran <quic_nmaran@quicinc.com>