ath11k: Fix unmapping allocated memory
In the present code we were not unmapping the allocated
memory, so in a scenario like where we do infinite number
of FW crash and recovery, we face a crash as we run out of
memory and not able to allocate memory for further mapping.

So, here we are cleaning up the old mappings while recovery is
happening.

Change-Id: Iaba31e41b116169697d04141a8427c1915ef70c0
Signed-off-by: Rahul Bhattacharjee <quic_rbhattac@quicinc.com>