ath11k: fix filling parameters of WMI TPC command
Currently, if TPE IE is not present as well as channel is not
a PSD channel, the tx power is taken as minimum of eirp power
and the regulatory TPE IE power. But, if TPE IE is not present,
regulatory TPE IE will be 0 and hence min will always return 0.

Add fix to use directly the eirp power when TPE IE is not present.

Fixes: fd76105ba ("ath11k: fill parameters for vdev_set_tpc_power wmi command")

Change-Id: Ic82993bc350ec092e4b6e7428627046f3a125a10
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>