ath11k: fix filling parameters of WMI TPC command
Currently, if TPE IE is not present as well as channel is not
a PSD channel, the tx power is taken as minimum of eirp power
and the regulatory TPE IE power. But, if TPE IE is not present,
regulatory TPE IE will be 0 and hence min will always return 0.

Add fix to use directly the eirp power when TPE IE is not present.

Fixes: fd76105ba ("ath11k: fill parameters for vdev_set_tpc_power wmi command")

Change-Id: I98b55775058e4dd2ce1cad4c9a93ec13cdb5edcf
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>