mac80211 : allow bw change during channel switch in mesh
From 'IEEE Std 802.11-2020 section 11.8.8.4' :
The mesh channel switch may be triggered by the need to avoid interference to a
detected radar signal, or to reassign mesh STA channels to ensure
the MBSS connectivity.
A 20/40 MHz MBSS may be changed to a 20 MHz MBSS and a 20 MHz MBSS
may be changed to a 20/40 MHz MBSS.

Since the standard allows the change of bandwidth during the channel switch in mesh,
remove the bandwidth check present in ieee80211_set_csa_beacon.

Fixes: c6da674aff94 ("{nl,cfg,mac}80211: enable the triggering of CSA frame in mesh")

Change-Id: I9edd318add059ca6800bfdb2271e0fb8bf848db8
Signed-off-by: Hari Chandrakanthan <quic_haric@quicinc.com>