ath11k: Fix proper skb handling in dynamic VLAN case
The HTT metadata and the respective alignment which are
added as the skb header for Dynamic VLAN case are not
removed at the TX completion handler before giving the
skb to mac80211 for status update.

So, when monitor interface is enabled, the skb buffers
which needs to sent to the monitor interface have
insufficient headroom and the packet drops.

Removed the HTT metadata and alignment from the skb buffer
at the TX completion handler if it is added during the
transmission.

Change-Id: Ib3a317e538b53e20751a63f99bf3c80a9a73f654
Signed-off-by: Nagarajan Maran <quic_nmaran@quicinc.com>