ath12k: add drop descriptor handling in monitor ring
In stretch test scenario, driver receives invalid values in the cookies
of rx monitor destination ring. But the cookies which has junk values
are corresponding to drop descriptor and hardware expects driver to avoid
handling or proccessing of this descriptor by advertising EMPTY_DESCRIPTOR
as true in the destination descriptor.

Skip the destination descriptor processing when the EMPTY_DESCRIPTOR bit
is set as true to avoid processing of junk cookies. Also mon_drop_desc
entry in soc_dp_stats captures the number of drop descriptor received from the
monitor destination ring.

Change-Id: I83a21e77c7632e9592766e3a56a5c490333fe5cf
Signed-off-by: P Praneesh <quic_ppranees@quicinc.com>