ath12k:dp Rx performance improvements
653-002: Disable CMD_RXMON_GCONF_EN in ath12k_dp_tx_htt_rx_filter_setup()
when stats is disabled. This prevents interrupts being generated for monitor
ring. This eventually improves the CPU idle by 15 to 18%.

661-001: Modified Ring size and buffer size for Rx DMA and Rx monitor.
Reduced Tx batch threshold to 64 and Tx timer threshold to 200.
This is to inline with prop driver and with modified ring size gains
Throughput buy 100 to 150 Mbps.

661-002: Introduced module param for enabling fwlog. By default, fw logging is
disabled. This eventually reduces the CE interrupts which helps in
gaining CPU IDLE by 2%.

Change-Id: I8462be549e69a0fc2844d8979382d9b9b52a043e
Signed-off-by: Balamurugan Selvarajan <quic_bselvara@quicinc.com>