ath11k/ath12k: fix clearing peer keys during sta state auth to assoc
During station state change from AUTHORIZED to ASSOC, drivers must clear
its key from hw. Ath11k clearing the keys during ASSOC to AUTH which is not a
valid sequence and there is a chance of accessing ptr after free.

Clear the peer keys while the state of station changes from AUTHORIZED
to ASSOC.

During peer cleanup in failure case of rx_tid, idx reused and it is starting
from tid 17 which is not a valid tid, so fixed it to clean from tid 16 to
0.

Added a sanity check with tid max value.

Change-Id: I2411cc5c6e999660cde9fe477861d645a1393e68
Signed-off-by: Karthikeyan Kathirvel <quic_kathirve@quicinc.com>