ath12k: Configure rx hash map for 2 GHz, 5 GHz, 6 GHz
Configure rx hash map for all three radio's.
When hash based routing of rx packet is enabled, 32 entries to map
the hash values to the ring will be configured. Each hash entry uses
four bits to map to a particular ring.
Currently the values configured, will limit to three rings. The values
differs for IX2 and IX3, So that interrupts will not be routed to the
same core. This will be fine tuned eventually.

Change-Id: I0eb3e831045565260b9d76a911ef2742554884fb
Signed-off-by: Balamurugan Selvarajan <quic_bselvara@quicinc.com>