ath12k: fill peer meta data during reo_reinject
Firmware expects physical address in the reo entrance ring when MLO is
enabled and peer meta data in the queue_addr_lo for the non MLO cases.
Current implementation fills the physical address for the non MLO cases
and firmware decodes it differently which leads to SOC CSR region access
which is not related to firmware/MAC operation.

Fix it by adding peer meta data in the reo entrance ring instead of physical
address since MLO is not yet supported.

Change-Id: I197a550af00a6daefd106b976c0911e3663b7ded
Signed-off-by: P Praneesh <quic_ppranees@quicinc.com>