ath12k: Update bank profiles during vif offload update
By default ath12k frame mode will be ETHERNET mode
The encap type will be sent to FW through WMI command
and bank profile will be configured for the vif.
When monitor vap is brought up, the frame mode for the vif
will be NATIVE WIFI. In this case encap type is sent as NATIVE
WIFI through WMI command but bank profile for the vif is not
updated . This difference is encap type causes transmit failures.

Updating the bank profile before sending WMI command
when there is a change in encap type.

Change-Id: I1af626da6e1af5fcf89f3194174a29d7d1568ffa
Signed-off-by: Ramya Gnanasekar <quic_rgnanase@quicinc.com>