ath12k: fix radiotap center frequency updation for 6G
Center frequency which is provided under HAL_RX_PPDU_START is the
starting frequency of the total bandwidth, whereas radio tap header
expects current operating channel frequency.

Fix it by converting channel number to channel frequency and assign
it to rx_status, so that it can be further used it for building
radiotap header.

Change-Id: Id1425719873d80d979c51232c65807cfc8c23a40
Signed-off-by: P Praneesh <quic_ppranees@quicinc.com>