ath12k: fix MSI total_vector count as power of 2
MSI total vector has value of 20 (12 DP + 5 CE + 3 MHI interrupts).
But the MSI specification only allows interrupts to be allocated in power of
two and up to a maximum of 2 power 5 which is 32. Since the total vector
carries count as non power of 2 value, MSIs are not freed properly during
rmmod which results in module load failure in the consecutive insmod
due to unavailability of MSIs in the host.

Fix it by changing the total vector count as 32 which is satisfies MSI
specification.

Change-Id: I9dba159f65c4175d83e546ef49fe2dd75379037a
Signed-off-by: P Praneesh <quic_ppranees@quicinc.com>