mac80211: update hostapd vht config based on vht capab
VHT capability is updated with VHT160 only when both 160MHz
and 80+80 bit is set in VHT phy capability information.

Updating VHT160 to be configured in hostapd when 160MHz
supported bit is set in phy capability information.
When 80+80 MHz supported bit is set VHT160-80PLUS80 will be
configured in hostapd configuration.

Change-Id: I30f05e8e719193d4b78229b197427d9d0768a1e6
Signed-off-by: Ramya Gnanasekar <quic_rgnanase@quicinc.com>