ath12k: [single wiphy] incorrect channel survey dump
For survey dump collection, freq_to_idx() api is used
to map the index of particular frequency to populate its
survey.

In case of single wiphy, wiphy is aware of all the
frequency bands present, the index was incorrect because
the mapping did not consider whether the freq band is
present in its radio structure (ar). Hence, channel survey
sent to userspace was invalid.

To handle this, the freq_to_idx() api should map index for the
particular band in the ar structure.

Change-Id: I17d197a71c955c96ba3d1dcedc2cf2786237a376
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>