ath12k: Fix freeing hal reo status entry in host in failure case
Host free up the memory in ring full scenario
for sending HAL_REO_CMD_*.

But hardware has the memory uncleared since
it might not received the command to flush. Due to this
hardware may over write some value on the memory
and that leads memory corruption in host.

Fix this by keeping the flush cache list intact in
failure case and retry later to avoid memory corruption.

Change-Id: I12b02ac8114f3d0cefcb7e8d624b0ff635deda3a
Signed-off-by: Manish Dharanenthiran <quic_mdharane@quicinc.com>